+++ src/main/java/com/paneedah/weaponlib/crafting/ammopress/TileEntityAmmoPress.java
@@ -56,7 +56,8 @@ public class TileEntityAmmoPress extends TileEntityStation {
 	
 	
 	public ItemStack getLatestStackInQueue() {
-		if(this.craftStack.isEmpty()) return null;
+		if(this.craftStack.isEmpty())
+			return ItemStack.EMPTY;
 		ItemStack stack = craftStack.peek();
 		if(stack.isEmpty()) {
 			craftStack.pop();
@@ -94,7 +95,7 @@ public class TileEntityAmmoPress extends TileEntityStation {
 	}
 	
 	public boolean hasStack() {
-		return !this.craftStack.isEmpty() && getLatestStackInQueue() != null;
+		return !this.craftStack.isEmpty() && getLatestStackInQueue() != ItemStack.EMPTY;
 	}
 	
 	public void addStack(ItemStack stack) {